pyproject.toml
[tool.poetry]
name = "goit-ds-hw-03"
version = "0.1.0"
description = ""
authors = ["Kagev <kaplitskiygv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest-playwright = "^0.5.2"
scrapy = "^2.10.0"
aioredis = "^2.0.1"
colorama = "^0.4.6"
pymongo = "^4.5.0"
playwright = "^1.47.0"
mongoengine = "^0.29.0"
pika = "^1.3.2"
python-dotenv = "^1.0.1"
lxml = "^5.3.0"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
